fix link scroll adjustment for ng and non-ng pages (#487)
Fixes #315
Fixes #483

With this fix, link anchor offsets get handled in CSS using the same
solution as that implemented for site-www. We also disable the scroll
adjustments that used to be done via the `scroll-y-offset-element`
directive (for angular doc pages). This way we have a uniform scheme
that works for all pages angular and non-angular pages.

The offsetting is achieved via an "anchor" class which is set for
header and anchor elements with id's. It is the `jekyll-toc` gem that
adds this class, but only for pages for which the gem's functionality
is enabled -- setting the Jekyll page property `toc: false` disables it.

Comments on individual changes:
- src/_assets/javascripts/main.js: remove on load function which
attempted to adjust the y scroll position to compensate for the main
nav. This failed to work on site-www and was commented out there too.
Here, in addition, it also clashed with the ng directive.
- src/_assets/stylesheets/main.scss: define `anchor` class; cf.
https://github.com/dart-lang/site-www/pull/154.
  - This also adds an octicon-link next to anchors or headers with ids
  - The supporting icon file is:
src/_assets/icons/ic_link_black_24px.svg
- src/resources/js/directives/scroll-y-offset-element.js: don't adjust
the y scroll offset for ng pages anymore. (Note that removing the
directive entirely prevents ng hash paths (`#!#`) from being processed.)