Replace homepage with a link to our github repo.
The core landing page for angulardart.dev does not mention Angular at all. This is actually more harmful than just having a blank page. 

This PR changes the "Getting Started" link to point to the github README, which actually has information about angular. It also removes the rest of the useless content.